refactor: organise contextmenu JS logic into folders
