+++ nodes/web/helpers/contextmenu/main.js
@@ -1,0 +1,3 @@
+import { addMenuHandler, addNode } from "./utils.js";
+
+export { addMenuHandler, addNode };
